Add files via upload
The gender pay gap is a measure of income disparity between men and women. It is commonly calculated by dividing women's wages by men's wages, and this ratio is often expressed as a percent, or in dollar terms.
Women are generally considered to be paid less than men, and the gender pay gap exists for every age group and tends to increase throughout a woman's work life.
To get better clarity on the gender pay gap, the dataset was gotten from glassdoor for this analysis.
Context
Analyzing the pay gap for different job titles
About the data set:
The data set originates from Glassdoor and focuses on gender-based income for various job titles. This data collection will be useful in determining the depth of the gender-based pay gap, as several studies have shown that women are paid less than men for the same job titles. The following are the features of the data set: 
1. Job Title
2. Gender
3. Age
4. Performance Evaluation (Performance evaluation score)
5. Education (i.e Level of education)
6. Department
7. Seniority (i.e Number of years of experience)
8. Base Pay (Annual Basic Salary in dollars)
9. Bonus (Annual Bonus Pay)